fix: refactor to split respect eslint rules
