+++ backend/libs/room-operations.ts
@@ -5,6 +5,7 @@ import {
     VideoSyncStatus,
     Watcher,
 } from '../../extension/src/types';
+import { marshallRoom, unmarshallRoom } from './room-marshalling';
 
 /**
  * Find a room by ID in DDB
@@ -120,101 +121,5 @@ export const joinExistingRoom = async (
     } catch (e) {
         console.error(`Failed to create a room on DDB ${e}`);
         return undefined;
-    }
-};
-
-/**
- * RoomForDDB --> Room
- */
-const unmarshallRoom = (roomDDB: DocumentClient.AttributeMap): Room => {
-    const room: Room = {
-        roomId: roomDDB.roomId || 'error',
-        createdAt: new Date(roomDDB.createdAt),
-        ownerId: roomDDB.ownerConnectionString,
-        watchers: unmarshallWatchers(roomDDB.watchers),
-        minBufferLength: roomDDB.minBufferLength,
-        videoSpeed: roomDDB.videoSpeed,
-        currentVideoUrl: roomDDB.currentVideoUrl,
-        syncStartedAt: roomDDB.syncStartedAt
-            ? new Date(roomDDB.syncStartedAt)
-            : undefined,
-        syncStartedTimestamp: roomDDB.syncStartedTimestamp
-            ? new Date(roomDDB.syncStartedTimestamp)
-            : undefined,
-        videoStatus: roomDDB.videoStatus.WAITING,
-        resumePlayingAt: roomDDB.resumePlayingAt
-            ? new Date(roomDDB.resumePlayingAt)
-            : null,
-        resumePlayingTimestamp: roomDDB.resumePlayingTimestamp
-            ? new Date(roomDDB.resumePlayingTimestamp)
-            : undefined,
-    };
-    // TODO : check casting was good / we do not have out of date DDB items
-    return room;
-};
-
-/**
- * WatcherFromDDB --> Watcher
- */
-const unmarshallWatchers = (
-    watchersDDB: DocumentClient.AttributeMap,
-): Record<string, Watcher> => {
-    return Object.fromEntries(
-        Object.entries(watchersDDB).map(([watcherID, watcherDDB]) => [
-            watcherID,
-            {
-                id: watcherDDB.id,
-                connectionId: watcherDDB.connectionId,
-                joinedAt: new Date(watcherDDB.joinedAt),
-                lastVideoTimestamp: watcherDDB.lastVideoTimestamp
-                    ? new Date(watcherDDB.lastVideoTimestamp)
-                    : undefined,
-                lastHeartbeat: new Date(watcherDDB.lastHeartbeat),
-                currentVideoStatus: watcherDDB.currentVideoStatus,
-                initialSync: watcherDDB.initialSync,
-                userAgent: watcherDDB.userAgent,
-            },
-        ]),
-    );
-};
-
-/**
- * Room --> RoomForDDB
- */
-const marshallRoom = (room: Room) => {
-    const roomForDDB = marshallMap(room);
-    roomForDDB.watchers = marshallWatchers(roomForDDB.watchers);
-    return roomForDDB;
-};
-
-/**
- * watchers --> watchersForDDB
- */
-const marshallWatchers = (watchers: Record<string, Watcher>) => {
-    return Object.fromEntries(
-        Object.entries(watchers).map(([watcherID, watcherDDB]) => [
-            watcherID,
-            marshallMap(watcherDDB),
-        ]),
-    );
-};
-
-/**
- * Converts a JS object to a compatible DDB format
- */
-const marshallMap = (map: Record<string, any>): Record<string, any> => {
-    return Object.fromEntries(
-        Object.entries(map).map(([key, value]) => [key, marshallValue(value)]),
-    );
-};
-
-/**
- * Dates are converted to strings (ISO 8601)
- */
-const marshallValue = (value: any): any => {
-    if (value instanceof Date) {
-        return value.toISOString;
-    } else {
-        return value;
     }
 };
